requirements.txt
requests
tqdm
python-dotenv
google-genai
opencv-python
tiktoken
replicate